+++ integration/teardown_unix.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+docker stop int-ssh-server
+docker rm int-ssh-server